requirements.txt
dspy-ai~=2.4.0
neo4j~=5.18.0
pyautogen~=0.2.23
openai~=1.17.1
pydantic~=2.5.0
datasets~=2.14.7
python-dotenv~=1.0.1
